backend/server/requirements.txt
requests
fastapi
psycopg2-binary
uvicorn
databases[postgresql]
